requirements.txt
numpy==1.22.3
pandas==1.4.3
performer-pytorch==1.1.4
scikit-learn==1.1.1
scipy==1.8.1
transformers==4.21.0